sd-dhcp-client: store lifetime and friends in usec_t

+++ src/systemd/sd-dhcp-client.h
@@ -141,7 +141,7 @@ int sd_dhcp_client_set_socket_priority(
                 int so_priority);
 int sd_dhcp_client_set_fallback_lease_lifetime(
                 sd_dhcp_client *client,
-                uint32_t fallback_lease_lifetime);
+                uint64_t fallback_lease_lifetime);
 
 int sd_dhcp_client_add_option(sd_dhcp_client *client, sd_dhcp_option *v);
 int sd_dhcp_client_add_vendor_option(sd_dhcp_client *client, sd_dhcp_option *v);
